refactor: GoogleCalendarAPIを認証・API・エラー型に責務分離 (#116)
487行のGoogleCalendarAPI.swiftを3ファイルに分割:
- GoogleCalendarAPI.swift (211行) - カレンダーCRUD操作
- GoogleAuthService.swift (223行) - OAuth認証・トークン管理
- CalendarAPIError.swift (32行) - 統一エラー型

主な改善点:
- 認証とAPI操作の責務を明確に分離
- CalendarAPIErrorで統一的なエラーハンドリング
- performRequest/validateResponseで重複コードを共通化
- 呼び出し元（SettingView, WorkoutRecordView等）を更新
