+++ TrackFit/Views/WorkoutRecordView/WorkoutRecordView.swift
@@ -357,14 +357,14 @@ struct WorkoutRecordView: View {
                     isShowCalendarIntegration = true
                 }
                 Task {
-                    await GoogleCalendarAPI.checkAndUpdateLinkingStatus()
+                    await GoogleAuthService.checkAndUpdateLinkingStatus()
                 }
             }
         }
         .onChange(of: scenePhase) { _, newPhase in
             if newPhase == .active && isCalendarFeatureEnabled {
                 Task {
-                    await GoogleCalendarAPI.checkAndUpdateLinkingStatus()
+                    await GoogleAuthService.checkAndUpdateLinkingStatus()
                 }
             }
         }
